+++ examples/drag_window.rs
@@ -28,8 +28,8 @@ fn main() -> Result<(), impl std::error::Error> {
         Event::WindowEvent { event, window_id } => match event {
             WindowEvent::CloseRequested => elwt.exit(),
             WindowEvent::MouseInput {
-                state: ElementState::Pressed,
-                button: MouseButton::Left,
+                state,
+                button,
                 ..
             } => {
                 let window = if (window_id == window_1.id() && switched)
@@ -40,7 +40,11 @@ fn main() -> Result<(), impl std::error::Error> {
                     &window_1
                 };
 
-                window.drag_window().unwrap()
+                match (button, state) {
+                    (MouseButton::Left, ElementState::Pressed) => window.drag_window().unwrap(),
+                    (MouseButton::Right, ElementState::Released) => window.show_window_menu(),
+                    _ => ()
+                }
             }
             WindowEvent::CursorEntered { .. } => {
                 entered_id = window_id;
@@ -54,10 +58,26 @@ fn main() -> Result<(), impl std::error::Error> {
                         ..
                     },
                 ..
-            } if c == "x" => {
-                switched = !switched;
-                name_windows(entered_id, switched, &window_1, &window_2);
-                println!("Switched!")
+            } => {
+                match c.as_str() {
+                    "x" => {
+                        switched = !switched;
+                        name_windows(entered_id, switched, &window_1, &window_2);
+                        println!("Switched!")
+                    },
+                    "d" => {
+                        let window = if (window_id == window_1.id() && switched)
+                            || (window_id == window_2.id() && !switched)
+                        {
+                            &window_2
+                        } else {
+                            &window_1
+                        };
+
+                        window.set_decorations(!window.is_decorated());
+                    }
+                    _ => ()
+                }
             }
             WindowEvent::RedrawRequested => {
                 if window_id == window_1.id() {