+++ src/platform_impl/android/mod.rs
@@ -933,7 +933,7 @@ impl Window {
     }
 
     #[inline]
-    pub fn show_system_menu(&self) {}
+    pub fn show_window_menu(&self) {}
 
     pub fn set_cursor_hittest(&self, _hittest: bool) -> Result<(), error::ExternalError> {
         Err(error::ExternalError::NotSupported(