+++ src/platform_impl/linux/mod.rs
@@ -440,7 +440,7 @@ impl Window {
     }
 
     #[inline]
-    pub fn show_system_menu(&self) {}
+    pub fn show_window_menu(&self) {}
 
     #[inline]
     pub fn set_cursor_hittest(&self, hittest: bool) -> Result<(), ExternalError> {